Alert user on install

+++ install.sh
@@ -1,5 +1,20 @@
 #!/usr/bin/env bash
+set -e
 DIR="$( cd "$( dirname "{BASH_SOURCE[0]}" )" && pwd )"
-rm ~/.vimrc
-ln -s $DIR/vimrc ~/.vimrc
-ln -sT $DIR ~/.vim
+
+read -p "This will REMOVE your .vimrc and .vim directory. Press Y to confirm." -n 1 -r
+echo # move to new line
+
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "Removing old installation"
+    rm ~/.vimrc
+    rm ~/.vim
+    echo "Creating symlinks"
+    ln -s $DIR/vimrc ~/.vimrc
+    ln -sT $DIR ~/.vim
+    echo "Done!"
+else
+    echo "Installation aborted"
+    exit 1
+fi
